Split nimby.rs (1410 lines) into nimby/ subdirectory (~5 files, all under 500 lines)
Refactor the monolithic nimby.rs into a subdirectory module:
- nimby/mod.rs (75 lines): re-exports + plugin definition
- nimby/types.rs (284 lines): constants, data types, Saveable impl, helpers
- nimby/opinion.rs (148 lines): pure opinion calculation functions
- nimby/systems.rs (309 lines): ECS systems
- nimby/tests.rs (396 lines): opinion/helper unit tests
- nimby/tests_state.rs (262 lines): state/saveable/personality unit tests

All public items are re-exported from mod.rs so external callers are unaffected.
